Added support for dynamic heading names, additional columns and code-cleanup
* Added support to show the card name specific to each row
If DRILL_HEADING is left blank (or, when prompted, pressing RET without typing anything),
the contents of the first column are used as the card name for that row.

* Added support for specifying additional columns
Most of the code to do this was already in place.
Only thing needed to be added was a check to see if the row had anything
specified in that column to prevent creating an empty card-subheading.

* Code-cleanup: updated org-drill-table--skip-props-and-schedule
If a heading didn't have any PROPERTIES or SCHEDULED, the point wouldn't move.
This is why we had to use (-drop 1) in org-drill-table--subtree->card