Fix for extra columns being left blank
Earlier if the 3rd or higher column was left blank then we wouldn't insert it as a card sub-heading.
eg. in the following table
;;  |-----------+---------+----------------|
;;  | English   | Spanish | Example        |
;;  |-----------+---------+----------------|
;;  | Today     | Hoy     | Hoy es domingo |
;;  | Yesterday | Ayer    |                |
;;  |-----------+---------+----------------|

we wouldn't have insert an Example heading for Yesterday.
However, it would have been included while creating the OrgDrillCard structure.
eg. (("English" . "Yesterday")("Spanish" . "Ayer")("Example" . ""))

This results in erroneous behavior if we run `org-drill-table-generate` again because the card that already exists would
not have the "Example" sub-heading and would result in a mismatch causing the "Yesterday" card to get created again.